requirements-keep.txt
gkeepapi==0.17.1
python-dotenv==1.2.2